+++ app/src/main/java/com/example/mycarsmt/model/repo/note/NoteService.kt
@@ -6,8 +6,8 @@ import com.example.mycarsmt.model.Note
 interface NoteService{
 
     fun create(note: Note): Long
-    fun update(note: Note): Long
-    fun delete(note: Note): Long
+    fun update(note: Note): Int
+    fun delete(note: Note): Int
     fun readById(id: Long): LiveData<Note>
     fun readAll(): LiveData<List<Note>>
