+++ app/src/main/java/com/example/mycarsmt/model/repo/repair/RepairService.kt
@@ -6,8 +6,8 @@ import com.example.mycarsmt.model.Repair
 interface RepairService{
 
     fun create(repair: Repair): Long
-    fun update(repair: Repair): Long
-    fun delete(repair: Repair): Long
+    fun update(repair: Repair): Int
+    fun delete(repair: Repair): Int
     fun readAll(): LiveData<List<Repair>>
     fun readById(id: Long): LiveData<Repair>
